+++ tests/testthat/test-utils-names.R
@@ -7,3 +7,39 @@ test_that("Utils names", {
     c("ESP", "POR", "RTA", "USA")
   ))
 })
+
+test_that("Problematic names", {
+  skip_on_cran()
+  skip_if_offline()
+
+  expect_snapshot(gb_helper_countrynames(c("Espagne", "Antartica")))
+  expect_snapshot(gb_helper_countrynames(c("spain", "antartica")))
+
+  ata <- get_gb_adm0("Antartica", simplified = TRUE)
+  expect_s3_class(ata, "sf")
+
+  # Special case for Kosovo
+  expect_snapshot(gb_helper_countrynames(c("Spain", "Kosovo", "Antartica")))
+  expect_snapshot(gb_helper_countrynames(c("ESP", "XKX", "DEU")))
+
+  kos <- get_gb_adm0("Kosovo", simplified = TRUE)
+  expect_s3_class(kos, "sf")
+
+  full <- get_gb_adm0(c("Antarctica", "Kosovo"), simplified = TRUE)
+  expect_s3_class(full, "sf")
+  expect_identical(full$shapeGroup, c("ATA", "XKX"))
+  expect_equal(nrow(full), 2)
+})
+
+test_that("Test full name conversion", {
+  skip_on_cran()
+  skip_if_offline()
+
+  allnames <- get_gb_meta(adm_lvl = "ADM0")
+  nm <- unique(allnames$boundaryName)
+  expect_silent(nm2 <- gb_helper_countrynames(nm))
+  isos <- unique(allnames$boundaryISO)
+  expect_silent(isos2 <- gb_helper_countrynames(isos))
+  expect_identical(length(nm), length(isos2))
+  expect_identical(length(nm), length(nm2))
+})