+++ NEWS.md
@@ -3,6 +3,7 @@
 -   Add DOI.
 -   `get_gb_cgaz()` get the latest data available on the repo
     <https://github.com/wmgeolab/geoBoundaries/tree/main/releaseData>.
+-   All functions: Improve detection for Antarctica and Kosovo.
 
 # geobounds 0.0.1